Move uri_parser to lib
