+++ gRPC-C++.podspec
@@ -274,7 +274,7 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/http/client/http_client_filter.h',
                       'src/core/ext/filters/http/message_compress/message_compress_filter.h',
                       'src/core/ext/filters/http/server/http_server_filter.h',
-                      'src/core/ext/uri/uri_parser.h',
+                      'src/core/lib/uri/uri_parser.h',
                       'src/core/ext/filters/client_channel/lb_policy/grpclb/grpclb.h',
                       'src/core/ext/filters/client_channel/lb_policy/xds/xds.h',
                       'src/core/lib/security/context/security_context.h',
