+++ gRPC-Core.podspec
@@ -275,7 +275,7 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/http/client/http_client_filter.h',
                       'src/core/ext/filters/http/message_compress/message_compress_filter.h',
                       'src/core/ext/filters/http/server/http_server_filter.h',
-                      'src/core/ext/uri/uri_parser.h',
+                      'src/core/lib/uri/uri_parser.h',
                       'src/core/ext/filters/client_channel/lb_policy/grpclb/grpclb.h',
                       'src/core/ext/filters/client_channel/lb_policy/xds/xds.h',
                       'src/core/lib/security/context/security_context.h',
@@ -703,7 +703,7 @@ Pod::Spec.new do |s|
                       'src/core/ext/filters/http/http_filters_plugin.cc',
                       'src/core/ext/filters/http/message_compress/message_compress_filter.cc',
                       'src/core/ext/filters/http/server/http_server_filter.cc',
-                      'src/core/ext/uri/uri_parser.cc',
+                      'src/core/lib/uri/uri_parser.cc',
                       'src/core/lib/http/httpcli_security_connector.cc',
                       'src/core/lib/security/context/security_context.cc',
                       'src/core/lib/security/credentials/alts/alts_credentials.cc',
@@ -894,7 +894,7 @@ Pod::Spec.new do |s|
                               'src/core/ext/filters/http/client/http_client_filter.h',
                               'src/core/ext/filters/http/message_compress/message_compress_filter.h',
                               'src/core/ext/filters/http/server/http_server_filter.h',
-                              'src/core/ext/uri/uri_parser.h',
+                              'src/core/lib/uri/uri_parser.h',
                               'src/core/ext/filters/client_channel/lb_policy/grpclb/grpclb.h',
                               'src/core/ext/filters/client_channel/lb_policy/xds/xds.h',
                               'src/core/lib/security/context/security_context.h',
