+++ grpc.gemspec
@@ -207,7 +207,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/http/client/http_client_filter.h )
   s.files += %w( src/core/ext/filters/http/message_compress/message_compress_filter.h )
   s.files += %w( src/core/ext/filters/http/server/http_server_filter.h )
-  s.files += %w( src/core/ext/uri/uri_parser.h )
+  s.files += %w( src/core/lib/uri/uri_parser.h )
   s.files += %w( src/core/ext/filters/client_channel/lb_policy/grpclb/grpclb.h )
   s.files += %w( src/core/ext/filters/client_channel/lb_policy/xds/xds.h )
   s.files += %w( src/core/lib/security/context/security_context.h )
@@ -639,7 +639,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/filters/http/http_filters_plugin.cc )
   s.files += %w( src/core/ext/filters/http/message_compress/message_compress_filter.cc )
   s.files += %w( src/core/ext/filters/http/server/http_server_filter.cc )
-  s.files += %w( src/core/ext/uri/uri_parser.cc )
+  s.files += %w( src/core/lib/uri/uri_parser.cc )
   s.files += %w( src/core/lib/http/httpcli_security_connector.cc )
   s.files += %w( src/core/lib/security/context/security_context.cc )
   s.files += %w( src/core/lib/security/credentials/alts/alts_credentials.cc )
