+++ src/core/ext/filters/client_channel/parse_address.h
@@ -23,8 +23,8 @@
 
 #include <stddef.h>
 
-#include "src/core/ext/uri/uri_parser.h"
 #include "src/core/lib/iomgr/resolve_address.h"
+#include "src/core/lib/uri/uri_parser.h"
 
 /** Populate \a resolved_addr from \a uri, whose path is expected to contain a
  * unix socket path. Returns true upon success. */
