+++ test/core/client_channel/uri_fuzzer_test.cc
@@ -23,8 +23,8 @@
 #include <grpc/grpc.h>
 #include <grpc/support/alloc.h>
 
-#include "src/core/ext/uri/uri_parser.h"
 #include "src/core/lib/iomgr/exec_ctx.h"
+#include "src/core/lib/uri/uri_parser.h"
 
 bool squelch = true;
 bool leak_check = true;
